Uses `request` package for npm installation
The `request` package allows the configuration of a proxy so Elm can
be installed behind proxies via npm.

The user has to set the environment variable "https_proxy" to the
proxy that is to be used.

Piping on the response object did not work anymore and had to be done
directly on the `request.get` call, therefore the tar and gunzip
functions were moved one level higher.

As a small fix, both error handlers at the start of the request
callback now return after rejecting the promise.